+++ src/MockPool.sol
@@ -46,7 +46,7 @@ contract MockPool is Initializable, OwnableUpgradeable, UUPSUpgradeable {
         _dispatcher = dispatcher;
     }
 
-    function dispatchAs(uint256 tokenId, DlnOrderLib.OrderCreation calldata data, uint32 _salt)
+    function dispatchAs(uint256 tokenId, DlnOrderLib.OrderCreation calldata data, uint64 _salt)
         external
         payable
         returns (uint256)
@@ -64,7 +64,7 @@ contract MockPool is Initializable, OwnableUpgradeable, UUPSUpgradeable {
         bondingToken.safeTransfer(owner(), amount);
     }
 
-    function _dispatch(uint256 tokenId, DlnOrderLib.OrderCreation calldata _orderCreation, uint32 _salt)
+    function _dispatch(uint256 tokenId, DlnOrderLib.OrderCreation calldata _orderCreation, uint64 _salt)
         internal
         returns (uint256)
     {
